requirements.txt
flask
sentence-transformers
faiss-cpu
pdfplumber
nltk
numpy
scikit-learn
groq
python-dotenv
